==> health/Health Care Services/Hospitals and Emergency Rooms/Service Access Scores/legacy/1.7.prepare_data_hosp_dmv_2021.Rmd <==
---
title: "prepare data for catchment_scores"
author: "--"
date: "2/4/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```
#libraries
```{r}
library(sf)
library(tidygeocoder)
library(dplyr)
library(community)
library(tidycensus)
```
#clean df
```{r}
#start clean
rm(list=ls())
```
#working directory
```{r}
setwd("~/VDH/Floating_Catchment_Areas/dmv/5.1 hospitals_dmv_model_2015_2022")
```
#obtain data. Do not run 
```{r}
# #a) conn
# library(RPostgreSQL)
# con <- RPostgreSQL::dbConnect(drv = RPostgreSQL::PostgreSQL(),
#                                dbname = "sdad",
#                                host = "postgis1",
#                                port = 5432,
#                                user = Sys.getenv(x = "DB_USR"),
#                                password = Sys.getenv(x = "DB_PWD"))
# 
# 
# #b) query
# provider <- sf::st_read(
#   con, 
#   query= "
# SELECT  id , name, address, county, countyfips, type, latitude, longitude, beds, geoid_blk, state
# FROM dc_health_behavior_diet.us_pl_hifld_2020_hospitals_geoblk A
# WHERE (state = 'VA' OR state = 'MD' OR state = 'DC') 
# AND  status='OPEN' 
# AND (type != 'PSYCHIATRIC' OR type != 'MILITARY') "
# )
# 
# #identify counties in NCR: National Capital Region
# counties_ncr <- c(51013 , 51059  ,51600  ,51107  ,51610  ,51683  ,51685  ,51153  ,51510  ,11001  , 24021  ,24031  ,24033  ,24017) 
# 
# provider <- provider %>% filter(as.numeric(countyfips) %in% c(counties_ncr))
# 
# #c) Disconnect
# RPostgreSQL::dbDisconnect(con)
# 
# #write
# write.csv(provider, "hospitals_ncr.csv" )
```

```{r}
#new data
provider_ncr <- read.csv("~/test_symlink/git/sdc.health_dev/Health Care Services/Hospitals and Emergency Rooms/Service Access Scores/data/working/ncr_cms_2015_2022_hospitals.csv")  
#only acute care hospitals
provider_ncr_acute <- provider_ncr %>% filter(hospital_type =='Acute Care Hospitals') 
#only year 2021
provider_ncr_acute_year <- provider_ncr %>% filter(hospital_type =='Acute Care Hospitals') %>% filter(year==2021)

```


# load data provider
```{r}
# provider <- read.csv("hospitals_ncr.csv", row.names = 1)
# provider <- provider %>% filter(type!= 'MILITARY')
# provider <- provider %>% filter(name!="NAVAL HEALTH CLINIC (JOHN HENRY BALCH)")

#assign name
provider <- provider_ncr_acute_year
#specify name
names(provider)[names(provider)=='long'] <- 'lon'
# assign IDs just to be explicit
provider$ID <- paste0("l", seq_len(nrow(provider)))
```

# data combined
```{r}
# ## data combined
# county_names <- list(
#   dc = "District of Columbia",
#   md = c("Charles", "Frederick", "Montgomery", "Prince George's"),
#   va = c("Alexandria", "Arlington", "Fairfax", "Falls Church", "Loudoun", "Manassas", "Manassas Park", "Prince William")
# )
# data <- list()
# shapes <- list()
# 
# 
# for(state in c("dc", "md", "va")){
#   # get county shapes to identify block groups
#   counties <- download_census_shapes(".", state, "county", paste0(state, "_counties"))
#   blockgroups <- download_census_shapes(".", state, "bg", paste0(state, "_blockgroups"))
#   
#   # save only selected counties
#   counties <- counties[counties$NAME %in% county_names[[state]],]
#   counties[counties$NAME == "Fairfax", "NAME"] <- c("Fairfax City", "Fairfax")
#   shapes[[state]] <- list(
#     counties = counties,
#     blockgroups = blockgroups[substr(blockgroups$GEOID, 1, 5) %in% counties$GEOID,]
#   )
#   
#   # population data
#   data[[state]] <- download_census_population(".", state, 2019, counties = counties$GEOID)
# }
# 
# data_combined <- do.call(rbind, lapply(names(data), function(state){
#   d <- data[[state]]$estimates
#   s <- shapes[[state]]$blockgroups
#   rownames(s) <- s$GEOID
#   total <- d$TOTAL.POPULATION_Total
#   total[total == 0] <- 1
#   data.frame(
#     GEOID = d$GEOID,
#     population = d$TOTAL.POPULATION_Total,
#     percent_female = d$SEX.BY.AGE_Female_Female / total * 100,
#     percent_white = d$RACE_Total_White.alone / total * 100,
#     percent_over_49 = rowSums(d[, grep("[5-8][05]", colnames(d))]) / total * 100,
#     female_over_14 = rowSums(d[, grep("Female_(?:[2-9]\\d|1[58])", colnames(d))]),
#     pediatric_pop = rowSums(d[, c(4:7,28:31) ]),
#     
#     st_coordinates(st_centroid(st_geometry(s[as.character(d$GEOID),])))
#   )
# }))
# data_combined <- data_combined[!is.na(data_combined$Y),]
```

# Data combined: population DMV 
```{r}
library(tidycensus)
library(tidyverse)

census_api_key("eba406410c653b81d6a795ac4e989221f7bdf302")

# counties_ncr <- c(51013 , 51059  ,51600  ,51107  ,51610  ,51683  ,51685  ,51153  ,51510  ,
#                   11001  , 
#                   24021  ,24031  ,24033  ,24017) 

buffer_blocks_ncr <- read.csv("~/VDH/Floating_Catchment_Areas/dmv/buffer_blocks_ncr.csv" )
buffer_blocks_ncr$fips <- substr(buffer_blocks_ncr$identifier,3,5)


# Bring in census tract data. 
pop_va_hosp <- get_acs(geography = "block group", 
                        year = 2021,
                        variables = c(population = "B01003_001E"),
                        state = "VA",
                       county = c('013', '059', '600', '107', '610', '683', '685', '153', '510'),
                       #county = c("013", "059") , 
                       survey = "acs5",
                        output = "wide",
                        geometry = TRUE)

pop_dc_hosp <- get_acs(geography = "block group", 
                        year = 2021,
                        variables = c(population = "B01003_001E"),
                        state = "DC",
                        survey = "acs5",
                        output = "wide",
                        geometry = TRUE)

pop_md_hosp <- get_acs(geography = "block group", 
                        year = 2021,
                        variables = c(population = "B01003_001E"),
                        state = "MD",
                       county = c('021', '031', '033', '017'),
                       survey = "acs5",
                        output = "wide",
                        geometry = TRUE)


#neighboring blocks MD: 24 - PA:42 - VA:51 - WV:54
pop_md_hosp_neig <- get_acs(geography = "block group", 
                        year = 2021,
                        variables = c(population = "B01003_001E"),
                        state = "MD",
                         county = unique(buffer_blocks_ncr$fips[buffer_blocks_ncr$STATEFP20=='24']),
                        survey = "acs5",
                        output = "wide",
                        geometry = TRUE)

pop_pa_hosp_neig <- get_acs(geography = "block group", 
                        year = 2021,
                        variables = c(population = "B01003_001E"),
                        state = "PA",
                         county = unique(buffer_blocks_ncr$fips[buffer_blocks_ncr$STATEFP20=='42']),
                        survey = "acs5",
                        output = "wide",
                        geometry = TRUE)

pop_va_hosp_neig <- get_acs(geography = "block group", 
                        year = 2021,
                        variables = c(population = "B01003_001E"),
                        state = "VA",
                         county = unique(buffer_blocks_ncr$fips[buffer_blocks_ncr$STATEFP20=='51']),
                        survey = "acs5",
                        output = "wide",
                        geometry = TRUE)

pop_wv_hosp_neig <- get_acs(geography = "block group", 
                        year = 2021,
                        variables = c(population = "B01003_001E"),
                        state = "WV",
                         county = unique(buffer_blocks_ncr$fips[buffer_blocks_ncr$STATEFP20=='54']),
                        survey = "acs5",
                        output = "wide",
                        geometry = TRUE)

#joint df
pop_buffer_hosp <- rbind(pop_va_hosp,
                         pop_dc_hosp,
                         pop_md_hosp,
                         pop_md_hosp_neig,
                         pop_pa_hosp_neig,
                         pop_va_hosp_neig,
                         pop_wv_hosp_neig ) %>% unique()


#
va <- pop_buffer_hosp %>% select(GEOID, NAME, population=population )


# data combined
#centroid and coordinates

data_combined <- data.frame(
  GEOID = va$GEOID,
  population = va$population,
sf::st_coordinates(st_centroid(va$geometry))
)

data_combined <- data_combined %>% filter(!is.na(data_combined$X))

data_combined <- data_combined %>% filter(GEOID != 517000323001 )
data_combined <- data_combined %>% filter(GEOID != 517000323002 )
data_combined <- data_combined %>% filter(GEOID != 517000323003 )

```


## travel time
```{r}
# library(osrm)
# options(osrm.server = Sys.getenv("OSRM_SERVER"), osrm.profile = "car")
# if(!file.exists("traveltimes_exercise.csv")){
#   traveltimes <- osrmTable(
#     src = data_combined[, c("GEOID", "X", "Y")],  #population-demand
#     dst = provider[, c("ID", "lon", "lat")]     #providers supply
#   )$duration
#   write.csv(
#     cbind(GEOID = rownames(traveltimes), as.data.frame(traveltimes)),
#     "traveltimes_exercise.csv", row.names = FALSE
#   )
# }
```

# Travel time
```{r}
library(osrm)
options(osrm.server = "http://104.248.112.16:5001/", osrm.profile = "car")

if(!file.exists("traveltimes_exercise.csv")){
  
  #provider <- st_drop_geometry( provider )
    
  traveltimes <- osrmTable(
    src = data_combined[, c( "X", "Y")],  #population-demand
    dst =  provider[, c( "lon", "lat")]  #providers supply
  )$duration
  
  names(traveltimes) <-   provider[, c("ID")] 
  
  write.csv(
    cbind(GEOID = data_combined[, c("GEOID")], as.data.frame(traveltimes)),
    "traveltimes_exercise.csv", row.names = FALSE
  )
}

traveltimes <- read.csv("traveltimes_exercise.csv", row.names = 1)
names(traveltimes) <-   provider[, c("ID")] 

```


# add1. Define geography id. This is because the Geography-GEOID from initial file may be outdated
```{r}
# library(tigris)
# library(maps)
# # add block geoids
# # get US blocks shapefile
# blocks_VA <- st_as_sf(blocks(state="VA"))
# blocks_MD <- st_as_sf(blocks(state="MD"))
# blocks_DC <- st_as_sf(blocks(state="DC"))
# blocks <- rbind(blocks_VA, blocks_MD)
# blocks <- rbind(blocks, blocks_DC)
# 
# # lon and lat to geo-points
# geopts <- provider %>%
#   st_as_sf(coords = c("lon", "lat"), crs = 4269)
# # indeces of bgs which contain a geopoint
# inds <- st_within(geopts$geometry, blocks$geometry, sparse=T)
# blk_list <- c()
# for (i in inds){
#   if (identical(blocks$NAME[i],character(0))){
#     blk_list<- append(blk_list, NA)}
#   else{
#     blk_list <- append(blk_list, blocks$GEOID[i])}
# }
# provider['geoid_blk_new'] <- blk_list
# 
# 
# # This assumes that you have longitude and latitude of each location and if you want bg or tract or county just replace 'blocks_VA <- st_as_sf(blocks(state="VA"))' with a corresponding value, e.g. 'tracts_VA <- st_as_sf(tracts(state="VA"))'. I believe it uses either tigris or maps package

```

# Define geography id - Geography-GEOID from initial file may be outdated
```{r}
library(tigris)
library(maps)
library(sf)
# add block geoids
# get US blocks shapefile
blocks_VA <- st_as_sf(block_groups(state="VA", year=2020), crs = 4326) #, year=2010
blocks_DC <- st_as_sf(block_groups(state= "DC", year=2020), crs = 4326) #, year=2010
blocks_MD <- st_as_sf(block_groups(state= "MD", year=2020), crs = 4326) #, year=2010
blocks_NC <- st_as_sf(block_groups(state= "NC", year=2020), crs = 4326) #, year=2010
blocks_DE <- st_as_sf(block_groups(state= "DE", year=2020), crs = 4326) #, year=2010
blocks_TN <- st_as_sf(block_groups(state= "TN", year=2020), crs = 4326) #, year=2010
blocks_WV <- st_as_sf(block_groups(state= "WV", year=2020), crs = 4326) #, year=2010
blocks_KY <- st_as_sf(block_groups(state= "KY", year=2020), crs = 4326) #, year=2010
blocks_PA <- st_as_sf(block_groups(state= "PA", year=2020), crs = 4326) #, year=2010

blocks <- rbind( blocks_VA, blocks_DC, blocks_MD , blocks_NC , blocks_DE , blocks_TN , blocks_WV , blocks_KY , blocks_PA) 
blocks <- st_transform(blocks, 4326)



# blocks_DC <- readRDS("~/VDH/Floating_Catchment_Areas/dmv/blocks_DC_2014" )
# blocks_MD <- readRDS( "~/VDH/Floating_Catchment_Areas/dmv/blocks_MD_2014" )
#blocks_VA <- readRDS("~/VDH/Floating_Catchment_Areas/dmv/blocks_VA_2014" )

# blocks <- rbind(blocks_VA, blocks_MD)
# blocks <- rbind(blocks, blocks_DC)

#blocks <- blocks_VA
# lon and lat to geo-points

geopts <- provider %>%
  st_as_sf(coords = c("lon", "lat"), crs = 4326)

geopts <- st_transform(geopts, 4326)

# indeces of bgs which contain a geopoint
inds <- st_within(geopts$geometry, blocks$geometry, sparse=T)
blk_list <- c()
for (i in inds){
  if (identical(blocks$NAME[i],character(0))){
    blk_list<- append(blk_list, NA)}
  else{
    blk_list <- append(blk_list, blocks$GEOID[i])}
}
provider['geoid_blk_new'] <- blk_list


# This assumes that you have longitude and latitude of each location and if you want bg or tract or county just replace 'blocks_VA <- st_as_sf(blocks(state="VA"))' with a corresponding value, e.g. 'tracts_VA <- st_as_sf(tracts(state="VA"))'. I believe it uses either tigris or maps package

```

# add2 count providers per geography using matching codes
```{r}
#obtain geoid

provider$GEOID <- substr(provider$geoid_blk_new, 1,12)
data_combined$GEOID <- as.character(data_combined$GEOID)

provider$num <- 1
#number of providers expanded
num_providers_wide <- full_join(provider %>% select(GEOID, num ), data_combined %>% select(GEOID), by = "GEOID" )

#number providers by GEOID
hosp_cnt <- num_providers_wide %>% group_by(GEOID) %>% na.omit(name) %>% summarise(hosp_cnt = n( ) )

#join hosp_cnt to geographies
data_combined <- data_combined %>% left_join(hosp_cnt, by= "GEOID")
```
# add3 mean and median of 10 nearest drive times 
```{r}
#mean of 10 nearest
top_mean <- function(x) {  
   mean(head(sort(x ), 10) ) }
#median of 10 nearest
top_median <- function(x) {  
   median(head(sort(x ), 10) ) }
#apply rowwise
traveltimes_near <- data.frame(near_10_mean=apply(traveltimes, 1, top_mean), 
                               near_10_median=apply(traveltimes, 1, top_median)) 
#rownames_to_column(traveltimes_near, var = "GEOID")
traveltimes_near$GEOID <- row.names(traveltimes_near) 
#join mean median traveltimes to geographies
data_combined <- data_combined %>% left_join(traveltimes_near, by= "GEOID")
```
#prepara data for save: SELECT relevant variables. 
```{r}
#raw traveltimes: traveltimes matrix already estimated and with colnames arranged 
traveltimes <- read.csv("traveltimes_exercise.csv", row.names = 1)
#population: always recheck relevant population: ie. for pediatrics: pop 0-17 years
population <- data_combined %>% select(GEOID, population, hosp_cnt, near_10_mean, near_10_median)
# realign travel times
names(traveltimes) <-   provider[, c("ID")] 
traveltimes <- traveltimes[as.character(population$GEOID), provider$ID]


```
# save new data
```{r}
write.csv(provider[, c("ID", "address", "lat", "lon")], "provider.csv", row.names = FALSE)
write.csv(cbind(GEOID = rownames(traveltimes), traveltimes), "traveltimes_trimmed.csv", row.names = FALSE)
write.csv(population, "population.csv", row.names = FALSE)
```
